requirements.txt

fastapi
uvicorn
requests
slowapi
tenacity
pytest
python-dotenv
celery
requests
asyncpg
python-jose
sentry-sdk
python-jose 
redis
asyncpg

